Hover tests for declarative macros
