+++ tests/e2e/hover/basic.rs
@@ -24,42 +24,6 @@ fn let_mut() {
 }
 
 #[test]
-fn println() {
-    test_transform_plain!(Hover,r#"
-    fn main() {
-        p<caret>rintln!("The value of x is: {}", x);
-    }
-    "#,@r#"
-    source_context = """
-        p<caret>rintln!("The value of x is: {}", x);
-    """
-    highlight = """
-        <sel>println</sel>!("The value of x is: {}", x);
-    """
-    popover = '''
-    ```cairo
-    println
-    ```
-    ---
-    Prints to the standard output, with a newline.
-    This macro uses the same syntax as `format!`, but writes to the standard output instead.
-
-    # Panics
-    Panics if any of the formatting of arguments fails.
-
-    # Examples
-    ```cairo
-    println!(); // Prints an empty line.
-    println!(\"hello\"); // Prints "hello".
-    let world: ByteArray = "world";
-    println!("hello {}", world_ba); // Prints "hello world".
-    println!("hello {world_ba}"); // Prints "hello world".
-    ```
-    '''
-    "#)
-}
-
-#[test]
 fn assign_lhs() {
     test_transform_plain!(Hover,r#"
     fn main() {
